[ Buiding Flask Env ]
This is just simple example written by zenith11311.
Purpose of this repository is just to share my teamates
how use flask and bootstrap.

To make the process easy, wrote scripts that install several tools
that are necessary to operate in isolated environment,
such as machines that are connected to internet but no sudo privilege.

Anyone can use the code and make change at your own branch as you like.
I might update some features, but please don't expect much about it.

Have fun with it! :D

* Building up flask env
    * script for building tools (python3.7, redis-stable venv-16.4.1)
    * web javascript srcs ( bootstrap, slick, jquery )
    * simple template & button events...

* HOWTO
    * on commandline
        * $ cd $WORKING_PATH
        * $ source init.env / . init.env
        * $ cd flask
        * $ source flask.env / . flask.env
        * $ python run.py
    * on browser
        * url : http://localhost:7003